raspberry-pi5/requirements.txt
types-serial~=3.5.0
pyserial~=3.5.0
smbus2~=0.5.0
